Refactored spectrometer widget code. Added functionality that allows replots of data with recalculation of integrals and fit curves
